+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 This file is for listing all the changes to this project
 
 ## v1.0.1 Release | Oct 10, 2023
-- Using drawShadow method instead of draw in `TextScreen`
+### Changed
+- Adding parameter for drawShadow in `TextScreen`
+- Using **drawInternal** method instead of **draw** in `TextScreen`
+- Changing ComponentItem.ComponentType to Class instead of Enum
 
 ## v1.0.0 Release | Oct 7, 2023
 - Initial release!